--- Chat/ChatViewController.cs
@@ -1,793 +1,0 @@
-﻿using BeatSaberMarkupLanguage.Attributes;
-using BeatSaberMarkupLanguage.FloatingScreen;
-using BeatSaberMarkupLanguage.ViewControllers;
-using EnhancedStreamChat.Graphics;
-using EnhancedStreamChat.Utilities;
-using System;
-using System.Collections;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
-using TMPro;
-using UnityEngine;
-using UnityEngine.TextCore;
-using UnityEngine.UI;
-using ChatCore;
-using ChatCore.Services.Twitch;
-using ChatCore.Models.Twitch;
-using ChatCore.Interfaces;
-using ChatCore.Services;
-using UnityEngine.SceneManagement;
-using BS_Utils.Utilities;
-using BeatSaberMarkupLanguage.Components.Settings;
-using BeatSaberMarkupLanguage.Parser;
-using VRUIControls;
-using System.Diagnostics;
-using BeatSaberMarkupLanguage;
-
-namespace EnhancedStreamChat.Chat
-{
-    [HotReload]
-    public class ChatViewController : BSMLAutomaticViewController
-    {
-        private EnhancedFontInfo _chatFont;
-        private Queue<EnhancedTextMeshProUGUIWithBackground> _activeChatMessages = new Queue<EnhancedTextMeshProUGUIWithBackground>();
-        private ObjectPool<EnhancedTextMeshProUGUIWithBackground> _textPool;
-        private FloatingScreen _chatScreen;
-        private GameObject _gameObject;
-        private ChatConfig _chatConfig;
-        private Material _chatMoverMaterial;
-        private bool _isInGame = false;
-        private string _fontPath = Path.Combine(Environment.CurrentDirectory, "Cache", "FontAssets");
-
-        private void Start()
-        {
-            _chatConfig = ChatConfig.instance;
-            StartCoroutine(WaitForFont());
-            SetupScreens();
-
-            if (_textPool == null)
-            {
-                _textPool = new ObjectPool<EnhancedTextMeshProUGUIWithBackground>(20,
-                    Constructor: () =>
-                    {
-                        var go = new GameObject();
-                        DontDestroyOnLoad(go);
-                        var msg = go.AddComponent<EnhancedTextMeshProUGUIWithBackground>();
-                        msg.Text.enableWordWrapping = true;
-                        msg.Text.FontInfo = _chatFont;
-                        msg.SubText.enableWordWrapping = true;
-                        msg.SubText.FontInfo = _chatFont;
-                        UpdateChatMessage(msg);
-                        return msg;
-                    },
-                    OnAlloc: (msg) =>
-                    {
-                        msg.gameObject.transform.SetParent(_chatContainer.transform, false);
-                        if(ReverseChatOrder)
-                        {
-                            msg.gameObject.transform.SetAsFirstSibling();
-                        }
-                    },
-                    OnFree: (msg) =>
-                    {
-                        try
-                        {
-                            msg.HighlightEnabled = false;
-                            msg.AccentEnabled = false;
-                            msg.SubTextEnabled = false;
-                            msg.Text.text = null;
-                            msg.Text.ChatMessage = null;
-                            msg.SubText.text = null;
-                            msg.SubText.ChatMessage = null;
-                            msg.gameObject.SetActive(false);
-                            msg.gameObject.transform.SetParent(rectTransform);
-                            msg.Text.ClearImages();
-                            msg.SubText.ClearImages();
-                        }
-                        catch (Exception ex)
-                        {
-                            Logger.log.Error($"An exception occurred while trying to free CustomText object. {ex.ToString()}");
-                        }
-                    }
-                );
-            }
-
-            BSEvents.menuSceneActive += BSEvents_menuSceneActive;
-            BSEvents.gameSceneActive += BSEvents_gameSceneActive;
-            ChatConfig.instance.OnConfigChanged += Instance_OnConfigUpdated;
-            UpdateChatUI();
-        }
-
-        protected override void DidDeactivate(DeactivationType deactivationType)
-        {
-            if (deactivationType == DeactivationType.NotRemovedFromHierarchy)
-            {
-                CleanupOldMessages(true);
-            }
-            base.DidDeactivate(deactivationType);
-        }
-
-        // TODO: eventually figure out a way to make this more modular incase we want to create multiple instances of the ChatViewController
-        private static ConcurrentQueue<IChatMessage> _backupMessageQueue = new ConcurrentQueue<IChatMessage>();
-        protected override void OnDestroy()
-        {
-            base.OnDestroy();
-            BSEvents.menuSceneActive -= BSEvents_menuSceneActive;
-            BSEvents.gameSceneActive -= BSEvents_gameSceneActive;
-            ChatConfig.instance.OnConfigChanged -= Instance_OnConfigUpdated;
-            foreach (var msg in _activeChatMessages)
-            {
-                if (msg.Text.ChatMessage != null)
-                {
-                    _backupMessageQueue.Enqueue(msg.Text.ChatMessage);
-                }
-                if (msg.SubText.ChatMessage != null)
-                {
-                    _backupMessageQueue.Enqueue(msg.SubText.ChatMessage);
-                }
-                Destroy(msg);
-            }
-            _activeChatMessages.Clear();
-            Destroy(_gameObject);
-            if (_textPool != null)
-            {
-                _textPool.Dispose();
-                _textPool = null;
-            }
-            if (_chatScreen != null)
-            {
-                Destroy(_chatScreen);
-                _chatScreen = null;
-            }
-            if(_chatFont != null)
-            {
-                Destroy(_chatFont.Font);
-                _chatFont = null;
-            }
-            if(_chatMoverMaterial != null)
-            {
-                Destroy(_chatMoverMaterial);
-                _chatMoverMaterial = null;
-            }
-            //if (_loadedAssets.Count > 0)
-            //{
-            //    foreach (var asset in _loadedAssets.Values)
-            //    {
-            //        asset.Unload(true);
-            //    }
-            //    _loadedAssets.Clear();
-            //}
-        }
-
-        [UIAction("#post-parse")]
-        private void PostParse()
-        {
-            // bg
-            _backgroundColorSetting.editButton.onClick.AddListener(HideSettings);
-            _backgroundColorSetting.modalColorPicker.cancelEvent += ShowSettings;
-            _backgroundColorSetting.CurrentColor = _chatConfig.BackgroundColor.ToColor();
-            // accent
-            _accentColorSetting.editButton.onClick.AddListener(HideSettings);
-            _accentColorSetting.modalColorPicker.cancelEvent += ShowSettings;
-            _accentColorSetting.CurrentColor = _chatConfig.AccentColor.ToColor();
-            // highlight
-            _highlightColorSetting.editButton.onClick.AddListener(HideSettings);
-            _highlightColorSetting.modalColorPicker.cancelEvent += ShowSettings;
-            _highlightColorSetting.CurrentColor = _chatConfig.HighlightColor.ToColor();
-            // ping
-            _pingColorSetting.editButton.onClick.AddListener(HideSettings);
-            _pingColorSetting.modalColorPicker.cancelEvent += ShowSettings;
-            _pingColorSetting.CurrentColor = _chatConfig.PingColor.ToColor();
-            // text
-            _textColorSetting.editButton.onClick.AddListener(HideSettings);
-            _textColorSetting.modalColorPicker.cancelEvent += ShowSettings;
-            _textColorSetting.CurrentColor = _chatConfig.TextColor.ToColor();
-
-            _chatContainer.AddComponent<ContentSizeFitter>().verticalFit = ContentSizeFitter.FitMode.PreferredSize;
-        }
-
-        [UIParams]
-        internal BSMLParserParams parserParams;
-
-        [UIComponent("background-color-setting")]
-        ColorSetting _backgroundColorSetting;
-
-        [UIComponent("accent-color-setting")]
-        ColorSetting _accentColorSetting;
-
-        [UIComponent("highlight-color-setting")]
-        ColorSetting _highlightColorSetting;
-
-        [UIComponent("ping-color-setting")]
-        ColorSetting _pingColorSetting;
-
-        [UIComponent("text-color-setting")]
-        ColorSetting _textColorSetting;
-
-        [UIObject("ChatContainer")]
-        GameObject _chatContainer;
-
-        private Color _accentColor;
-        [UIValue("accent-color")]
-        public Color AccentColor
-        {
-            get => _accentColor;
-            set
-            {
-                _accentColor = value;
-                _chatConfig.AccentColor = "#" + ColorUtility.ToHtmlStringRGBA(value);
-                UpdateChatMessages();
-                NotifyPropertyChanged();
-            }
-        }
-
-        private Color _highlightColor;
-        [UIValue("highlight-color")]
-        public Color HighlightColor
-        {
-            get => _highlightColor;
-            set
-            {
-                _highlightColor = value;
-                _chatConfig.HighlightColor = "#" + ColorUtility.ToHtmlStringRGBA(value);
-                UpdateChatMessages();
-                NotifyPropertyChanged();
-            }
-        }
-
-        private Color _pingColor;
-        [UIValue("ping-color")]
-        public Color PingColor
-        {
-            get => _pingColor;
-            set
-            {
-                _pingColor = value;
-                _chatConfig.PingColor = "#" + ColorUtility.ToHtmlStringRGBA(value);
-                UpdateChatMessages();
-                NotifyPropertyChanged();
-            }
-        }
-
-        private Color _backgroundColor;
-        [UIValue("background-color")]
-        public Color BackgroundColor
-        {
-            get => _backgroundColor;
-            set
-            {
-                _backgroundColor = value;
-                _chatConfig.BackgroundColor = "#" + ColorUtility.ToHtmlStringRGBA(value);
-                _chatScreen.gameObject.GetComponent<Image>().material.color = value;
-                NotifyPropertyChanged();
-            }
-        }
-
-        private Color _textColor;
-        [UIValue("text-color")]
-        public Color TextColor
-        {
-            get => _textColor;
-            set
-            {
-                _textColor = value;
-                _chatConfig.TextColor = "#" + ColorUtility.ToHtmlStringRGBA(value);
-                UpdateChatMessages();
-                NotifyPropertyChanged();
-            }
-        }
-
-        [UIValue("font-size")]
-        public float FontSize
-        {
-            get => _chatConfig.FontSize;
-            set
-            {
-                _chatConfig.FontSize = value;
-                UpdateChatMessages();
-                NotifyPropertyChanged();
-            }
-        }
-
-        private int _settingsWidth = 110;
-        [UIValue("settings-width")]
-        public int SettingsWidth
-        {
-            get => _settingsWidth;
-            set
-            {
-                _settingsWidth = value;
-                NotifyPropertyChanged();
-            }
-        }
-
-        [UIValue("chat-width")]
-        public int ChatWidth
-        {
-            get => _chatConfig.ChatWidth;
-            set
-            {
-                _chatConfig.ChatWidth = value;
-                _chatScreen.ScreenSize = new Vector2(ChatWidth, ChatHeight);
-                NotifyPropertyChanged();
-            }
-        }
-
-        [UIValue("chat-height")]
-        public int ChatHeight
-        {
-            get => _chatConfig.ChatHeight;
-            set
-            {
-                _chatConfig.ChatHeight = value;
-                _chatScreen.ScreenSize = new Vector2(ChatWidth, ChatHeight);
-                NotifyPropertyChanged();
-            }
-        }
-
-        [UIValue("chat-position")]
-        public Vector3 ChatPosition
-        {
-            get => _isInGame ? _chatConfig.Song_ChatPosition : _chatConfig.Menu_ChatPosition;
-            set
-            {
-                if (_isInGame)
-                {
-                    _chatConfig.Song_ChatPosition = value;
-                }
-                else
-                {
-                    _chatConfig.Menu_ChatPosition = value;
-                }
-                _chatScreen.ScreenPosition = value;
-                NotifyPropertyChanged();
-            }
-        }
-
-        [UIValue("chat-rotation")]
-        public Vector3 ChatRotation
-        {
-            get => _isInGame ? _chatConfig.Song_ChatRotation : _chatConfig.Menu_ChatRotation;
-            set
-            {
-                if (_isInGame)
-                {
-                    _chatConfig.Song_ChatRotation = value;
-                }
-                else
-                {
-                    _chatConfig.Menu_ChatRotation = value;
-                }
-                _chatScreen.ScreenRotation = Quaternion.Euler(value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        [UIValue("allow-movement")]
-        public bool AllowMovement
-        {
-            get => _chatConfig.AllowMovement;
-            set
-            {
-                _chatConfig.AllowMovement = value;
-                _chatScreen.ShowHandle = value;
-                NotifyPropertyChanged();
-            }
-        }
-
-        [UIValue("chat-anchor-min-y")] 
-        public float ChatAnchorMinY
-        {
-            get => 0.022f;
-        }
-        [UIValue("chat-anchor-max-y")] 
-        public float ChatAnchorMaxY
-        {
-            get => 0.978f;
-        }
-        [UIValue("chat-pivot-pos-y")]
-        public float ChatPivotPosY
-        {
-            get => ReverseChatOrder ? 1.0f : 0.0f;
-        }
-        [UIValue("chat-anchor-pos-y")]
-        public float ChatAnchorPosY
-        {
-            get => ReverseChatOrder ? ChatAnchorMaxY : ChatAnchorMinY;
-        }
-        [UIValue("reverse-chat-order")]
-        public bool ReverseChatOrder
-        {
-            get => _chatConfig.ReverseChatOrder;
-            set
-            {
-                if(_chatConfig.ReverseChatOrder != value)
-                {
-                    // Reverse the order of any messages that are currently in chat
-                    var array = _chatContainer.GetComponentsInChildren<EnhancedTextMeshProUGUIWithBackground>().Reverse().ToArray();
-                    for (int i = 0; i < array.Length; i++)
-                    {
-                        array[i].gameObject.transform.SetSiblingIndex(i);   
-                    }
-                    _chatConfig.ReverseChatOrder = value;
-                }
-                NotifyPropertyChanged(nameof(ChatPivotPosY));
-                NotifyPropertyChanged(nameof(ChatAnchorPosY));
-                NotifyPropertyChanged(nameof(ChatAnchorMinY));
-                NotifyPropertyChanged(nameof(ChatAnchorMaxY));
-                NotifyPropertyChanged();
-            }
-        }
-
-        [UIValue("mod-version")]
-
-        public string ModVersion
-        {
-            get => Plugin.Version;
-        }
-
-        [UIAction("launch-web-app")]
-        private void LaunchWebApp()
-        {
-            ChatManager.instance._sc.LaunchWebApp();
-        }
-
-        [UIAction("launch-kofi")]
-        private void LaunchKofi()
-        {
-            Process.Start("https://ko-fi.com/brian91292");
-        }
-
-        [UIAction("launch-github")]
-        private void LaunchGitHub()
-        {
-            Process.Start("https://github.com/brian91292/EnhancedStreamChat-v3");
-        }
-
-        [UIAction("on-settings-clicked")]
-        private void OnSettingsClick()
-        {
-            Logger.log.Info("Settings clicked!");
-        }
-
-        [UIAction("#hide-settings")]
-        private void OnHideSettings()
-        {
-            Logger.log.Info("Saving settings!");
-            _chatConfig.Save();
-        }
-
-        private void HideSettings()
-        {
-            parserParams.EmitEvent("hide-settings");
-        }
-
-        private void ShowSettings()
-        {
-            parserParams.EmitEvent("show-settings");
-        }
-
-        private void BSEvents_gameSceneActive()
-        {
-            _isInGame = true;
-            AddToVRPointer();
-            UpdateChatUI();
-        }
-
-        private void BSEvents_menuSceneActive()
-        {
-            _isInGame = false;
-            AddToVRPointer();
-            UpdateChatUI();
-        }
-
-        private void AddToVRPointer()
-        {
-            VRPointer pointer = null;
-            if (_isInGame)
-            {
-                pointer = Resources.FindObjectsOfTypeAll<VRPointer>().Last();
-            }
-            else
-            {
-                pointer = Resources.FindObjectsOfTypeAll<VRPointer>().First();
-            }
-            if (_chatScreen.screenMover != null)
-            {
-                DestroyImmediate(_chatScreen.screenMover);
-                _chatScreen.screenMover = pointer.gameObject.AddComponent<FloatingScreenMoverPointer>();
-                _chatScreen.screenMover.Init(_chatScreen);
-                _chatScreen.screenMover.OnRelease += floatingScreen_OnRelease;
-                _chatScreen.screenMover.transform.SetAsFirstSibling();
-            }
-        }
-
-        private void Instance_OnConfigUpdated(ChatConfig config)
-        {
-            MainThreadInvoker.Invoke(() =>
-            {
-                UpdateChatUI();
-            });
-        }
-
-        private void SetupScreens()
-        {
-            if (_chatScreen == null)
-            {
-                _chatScreen = FloatingScreen.CreateFloatingScreen(new Vector2(ChatWidth, ChatHeight), true, ChatPosition, Quaternion.Euler(ChatRotation));
-                var canvas = _chatScreen.GetComponent<Canvas>();
-                canvas.sortingOrder = 3;
-                _chatScreen.SetRootViewController(this, true);
-                _gameObject = new GameObject();
-                DontDestroyOnLoad(_gameObject);
-                _chatMoverMaterial = Instantiate(BeatSaberUtils.UINoGlow);
-                _chatMoverMaterial.color = Color.clear;
-                var renderer = _chatScreen.handle.gameObject.GetComponent<Renderer>();
-                renderer.material = _chatMoverMaterial;
-                _chatScreen.transform.SetParent(_gameObject.transform);
-                var bg = _chatScreen.gameObject.GetComponent<UnityEngine.UI.Image>();
-                bg.material = Instantiate(BeatSaberUtils.UINoGlow);
-                AddToVRPointer();
-            }
-        }
-
-        private void floatingScreen_OnRelease(Vector3 pos, Quaternion rot)
-        {
-            ChatPosition = pos;
-            ChatRotation = rot.eulerAngles;
-            _chatConfig.Save();
-        }
-
-        private string BuildClearedMessage(EnhancedTextMeshProUGUI msg)
-        {
-            StringBuilder sb = new StringBuilder($"<color={msg.ChatMessage.Sender.Color}>{msg.ChatMessage.Sender.Name}</color>");
-            var badgeEndIndex = msg.text.IndexOf("<color=");
-            if (badgeEndIndex != -1)
-            {
-                sb.Insert(0, msg.text.Substring(0, badgeEndIndex));
-            }
-            sb.Append(": <color=#bbbbbbbb><message deleted></color>");
-            return sb.ToString();
-        }
-
-        private void ClearMessage(EnhancedTextMeshProUGUIWithBackground msg)
-        {
-            // Only clear non-system messages
-            if (!msg.Text.ChatMessage.IsSystemMessage)
-            {
-                msg.Text.text = BuildClearedMessage(msg.Text);
-                msg.SubTextEnabled = false;
-            }
-            if (msg.SubText.ChatMessage != null && !msg.SubText.ChatMessage.IsSystemMessage)
-            {
-                msg.SubText.text = BuildClearedMessage(msg.SubText);
-            }
-        }
-
-        private void UpdateChatUI()
-        {
-            rectTransform.localRotation = Quaternion.identity;
-            ChatWidth = _chatConfig.ChatWidth;
-            ChatHeight = _chatConfig.ChatHeight;
-            FontSize = _chatConfig.FontSize;
-            AccentColor = _chatConfig.AccentColor.ToColor();
-            HighlightColor = _chatConfig.HighlightColor.ToColor();
-            BackgroundColor = _chatConfig.BackgroundColor.ToColor();
-            PingColor = _chatConfig.PingColor.ToColor();
-            TextColor = _chatConfig.TextColor.ToColor();
-            ReverseChatOrder = _chatConfig.ReverseChatOrder;
-            if (_isInGame)
-            {
-                ChatPosition = _chatConfig.Song_ChatPosition;
-                ChatRotation = _chatConfig.Song_ChatRotation;
-            }
-            else
-            {
-                ChatPosition = _chatConfig.Menu_ChatPosition;
-                ChatRotation = _chatConfig.Menu_ChatRotation;
-            }
-            _chatScreen.handle.transform.localScale = new Vector2(ChatWidth, ChatHeight);
-            _chatScreen.handle.transform.localPosition = Vector3.zero;
-            _chatScreen.handle.transform.localRotation = Quaternion.identity;
-            AllowMovement = _chatConfig.AllowMovement;
-            UpdateChatMessages();
-        }
-
-        private void UpdateChatMessages()
-        {
-            foreach (var msg in _activeChatMessages)
-            {
-                UpdateChatMessage(msg, true);
-            }
-        }
-
-        private void UpdateChatMessage(EnhancedTextMeshProUGUIWithBackground msg, bool setAllDirty = false)
-        {
-            msg.Text.font = _chatFont.Font;
-            msg.Text.overflowMode = TextOverflowModes.Overflow;
-            msg.Text.alignment = TextAlignmentOptions.BottomLeft;
-            msg.Text.lineSpacing = 1.5f;
-            msg.Text.color = TextColor;
-            msg.Text.fontSize = _chatConfig.FontSize;
-            msg.Text.lineSpacing = 1.5f;
-
-            msg.SubText.font = _chatFont.Font;
-            msg.SubText.overflowMode = TextOverflowModes.Overflow;
-            msg.SubText.alignment = TextAlignmentOptions.BottomLeft;
-            msg.SubText.color = TextColor;
-            msg.SubText.fontSize = _chatConfig.FontSize;
-            msg.SubText.lineSpacing = 1.5f;
-
-            if (msg.Text.ChatMessage != null)
-            {
-                msg.HighlightColor = msg.Text.ChatMessage.IsPing ? PingColor : HighlightColor;
-                msg.AccentColor = AccentColor;
-                msg.HighlightEnabled = msg.Text.ChatMessage.IsHighlighted || msg.Text.ChatMessage.IsPing;
-                msg.AccentEnabled = !msg.Text.ChatMessage.IsPing && (msg.HighlightEnabled || msg.SubText.ChatMessage != null);
-            }
-
-            if (setAllDirty)
-            {
-                msg.Text.SetAllDirty();
-                if (msg.SubTextEnabled)
-                {
-                    msg.SubText.SetAllDirty();
-                }
-            }
-        }
-
-        private void CleanupOldMessages(bool force = false)
-        {
-            while (_activeChatMessages.TryPeek(out var nextClear) && (force || nextClear.transform.localPosition.y > _chatConfig.ChatHeight + 100))
-            {
-                _textPool.Free(_activeChatMessages.Dequeue());
-                //Logger.log.Info($"{_messageClearQueue.Count} messages shown");
-            }
-        }
-
-        public void OnMessageCleared(string messageId)
-        {
-            if (messageId != null)
-            {
-                MainThreadInvoker.Invoke(() =>
-                {
-                    foreach (var msg in _activeChatMessages)
-                    {
-                        if (msg.Text.ChatMessage == null)
-                        {
-                            continue;
-                        }
-                        if (msg.Text.ChatMessage.Id == messageId)
-                        {
-                            ClearMessage(msg);
-                        }
-                    }
-                });
-            }
-        }
-
-        public void OnChatCleared(string userId)
-        {
-            MainThreadInvoker.Invoke(() =>
-            {
-                foreach (var msg in _activeChatMessages)
-                {
-                    if (msg.Text.ChatMessage == null)
-                    {
-                        continue;
-                    }
-                    if (userId == null || msg.Text.ChatMessage.Sender.Id == userId)
-                    {
-                        ClearMessage(msg);
-                    }
-                }
-            });
-        }
-
-        public void OnJoinChannel(IChatService svc, IChatChannel channel)
-        {
-            MainThreadInvoker.Invoke(() =>
-            {
-                var newMsg = _textPool.Alloc();
-                newMsg.Text.text = $"<color=#bbbbbbbb>[{svc.DisplayName}] Success joining {channel.Id}</color>";
-                newMsg.HighlightEnabled = true;
-                newMsg.HighlightColor = Color.gray.ColorWithAlpha(0.05f);
-                newMsg.gameObject.SetActive(true);
-                _activeChatMessages.Enqueue(newMsg);
-
-                UpdateChatMessage(newMsg);
-                CleanupOldMessages();
-            });
-        }
-
-        public void OnChannelResourceDataCached(IChatChannel channel, Dictionary<string, IChatResourceData> resources)
-        {
-            MainThreadInvoker.Invoke(() =>
-            {
-                int count = 0;
-                foreach (var emote in resources)
-                {
-                    if (emote.Value.IsAnimated)
-                    {
-                        //Logger.log.Info($"Caching animated emote {emote.Key}");
-                        StartCoroutine(ChatImageProvider.instance.DownloadImage(emote.Value.Uri, "Emote", emote.Key, true));
-                        count++;
-                    }
-                }
-                Logger.log.Info($"{count} animated emotes total.");
-            });
-        }
-
-        EnhancedTextMeshProUGUIWithBackground _lastMessage;
-        public async void OnTextMessageReceived(IChatMessage msg)
-        {
-            if (_chatFont is null)
-            {
-                // TODO: maybe queue this message up or wait?
-                return;
-            }
-
-            string parsedMessage = await ChatMessageBuilder.BuildMessage(msg, _chatFont);
-
-            MainThreadInvoker.Invoke(() =>
-            {
-                if (_lastMessage != null && !msg.IsSystemMessage && _lastMessage.Text.ChatMessage.Id == msg.Id)
-                {
-                    // If the last message received had the same id and isn't a system message, then this was a sub-message of the original and may need to be highlighted along with the original message
-                    _lastMessage.SubText.text = parsedMessage;
-                    _lastMessage.SubText.ChatMessage = msg;
-                    _lastMessage.SubTextEnabled = true;
-                }
-                else
-                {
-                    var newMsg = _textPool.Alloc();
-                    newMsg.Text.ChatMessage = msg;
-                    newMsg.Text.text = parsedMessage;
-                    newMsg.gameObject.SetActive(true);
-                    _activeChatMessages.Enqueue(newMsg);
-                    _lastMessage = newMsg;
-                }
-                UpdateChatMessage(_lastMessage);
-                CleanupOldMessages();
-            });
-        }
-
-        private IEnumerator WaitForFont()
-        {
-            if(_chatFont != null)
-            {
-                yield break;
-            }
-
-           // yield return new WaitUntil(() => FontManager.IsInitialized);
-
-            TMP_FontAsset font = null;
-            string fontName = _chatConfig.SystemFontName;
-            if (!FontManager.TryGetTMPFontByFamily(fontName, out font))
-            {
-                Logger.log.Error($"Could not find font {fontName}! Falling back to Segoe UI");
-                fontName = "Segoe UI";
-            }
-            _chatFont = new EnhancedFontInfo(BeatSaberUtils.SetupFont(font));
-
-            foreach (var msg in _activeChatMessages)
-            {
-                msg.Text.SetAllDirty();
-                if (msg.SubTextEnabled)
-                {
-                    msg.SubText.SetAllDirty();
-                }
-            }
-
-            while (_backupMessageQueue.TryDequeue(out var msg))
-            {
-                OnTextMessageReceived(msg);
-            }
-        }
-    }
-}
