+++ Graphics/EnhancedTextMeshProUGUIWithBackground.cs
@@ -14,6 +14,8 @@ namespace EnhancedStreamChat.Graphics
         public EnhancedTextMeshProUGUI Text { get; internal set; }
         public EnhancedTextMeshProUGUI SubText { get; internal set; }
 
+        public event Action OnRenderRebuildComplete;
+
         private Image _highlight, _accent;
         private VerticalLayoutGroup _verticalLayoutGroup;
         public Vector2 Size
@@ -154,6 +156,7 @@ namespace EnhancedStreamChat.Graphics
         private void Text_OnLatePreRenderRebuildComplete()
         {
             (_accent.gameObject.transform as RectTransform).sizeDelta = new Vector2(1, (transform as RectTransform).sizeDelta.y);
+            OnRenderRebuildComplete?.Invoke();
         }
     }
 }
